replaced .unique with Unique()

+++ src/test/scala/no/mesan/testdatagen/generators/misc/FibonaccisSuite.scala
@@ -43,6 +43,6 @@ class FibonaccisSuite extends FunSuite with Printer {
 
   test("not two") {
     val expected= List(1, 1, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, 987, 1597, 2584)
-    assert(Fibonaccis().filter(n=> !(n equals 2)).get(expected.size) === expected)
+    assert(Fibonaccis().filter(n=> !(n.toInt equals 2)).get(expected.size) === expected)
   }
 }
